split the substitution formatter to different type
