+++ source/common/formatter/substitution_formatter.h
@@ -12,25 +12,13 @@
 #include "envoy/stream_info/stream_info.h"
 
 #include "source/common/common/utility.h"
+#include "source/common/formatter/stream_info_formatter.h"
 
 #include "absl/container/flat_hash_map.h"
 #include "absl/types/optional.h"
 
 namespace Envoy {
 namespace Formatter {
-
-class CommandSyntaxChecker {
-public:
-  using CommandSyntaxFlags = std::bitset<4>;
-  static constexpr CommandSyntaxFlags COMMAND_ONLY = 0;
-  static constexpr CommandSyntaxFlags PARAMS_REQUIRED = 1 << 0;
-  static constexpr CommandSyntaxFlags PARAMS_OPTIONAL = 1 << 1;
-  static constexpr CommandSyntaxFlags LENGTH_ALLOWED = 1 << 2;
-
-  static void verifySyntax(CommandSyntaxChecker::CommandSyntaxFlags flags,
-                           const std::string& command, const std::string& subcommand,
-                           const absl::optional<size_t>& length);
-};
 
 /**
  * Access log format parser.
@@ -40,78 +28,6 @@ public:
   static std::vector<FormatterProviderPtr> parse(const std::string& format);
   static std::vector<FormatterProviderPtr>
   parse(const std::string& format, const std::vector<CommandParserPtr>& command_parsers);
-
-  /**
-   * Parse a header subcommand of the form: X?Y .
-   * Will populate a main_header and an optional alternative header if specified.
-   * See doc:
-   * https://envoyproxy.io/docs/envoy/latest/configuration/observability/access_log/access_log#format-rules
-   */
-  static void parseSubcommandHeaders(const std::string& subcommand, std::string& main_header,
-                                     std::string& alternative_header);
-
-  /* Variadic function template to parse the
-     subcommand and assign found tokens to sequence of params.
-     subcommand must be a sequence
-     of tokens separated by the same separator, like: header1:header2 or header1?header2?header3.
-     params must be a sequence of std::string& with optional container storing std::string. Here are
-     examples of params:
-     - std::string& token1
-     - std::string& token1, std::string& token2
-     - std::string& token1, std::string& token2, std::vector<std::string>& remaining
-
-     If command contains more tokens than number of passed params, unassigned tokens will be
-     ignored. If command contains less tokens than number of passed params, some params will be left
-     untouched.
-  */
-  template <typename... Tokens>
-  static void parseSubcommand(const std::string& subcommand, const char separator,
-                              Tokens&&... params) {
-    std::vector<absl::string_view> tokens;
-    tokens = absl::StrSplit(subcommand, separator);
-    std::vector<absl::string_view>::iterator it = tokens.begin();
-    (
-        [&](auto& param) {
-          if (it != tokens.end()) {
-            if constexpr (std::is_same_v<typename std::remove_reference<decltype(param)>::type,
-                                         std::string>) {
-              // Compile time handler for std::string.
-              param = std::string(*it);
-              it++;
-            } else {
-              // Compile time handler for container type. It will catch all remaining tokens and
-              // move iterator to the end.
-              do {
-                param.push_back(std::string(*it));
-                it++;
-              } while (it != tokens.end());
-            }
-          }
-        }(params),
-        ...);
-  }
-
-  /**
-   * Return a FormatterProviderPtr if a built-in command is found. This method
-   * handles mapping the command name to an appropriate formatter.
-   *
-   * @param command - formatter's name.
-   * @param subcommand - optional formatter specific data.
-   * @param length - optional max length of output produced by the formatter.
-   * @return FormattterProviderPtr substitution provider for the command or nullptr
-   */
-  static FormatterProviderPtr parseBuiltinCommand(const std::string& command,
-                                                  const std::string& subcommand,
-                                                  absl::optional<size_t>& length);
-
-private:
-  using FormatterProviderCreateFunc =
-      std::function<FormatterProviderPtr(const std::string&, absl::optional<size_t>&)>;
-
-  using FormatterProviderLookupTbl =
-      absl::flat_hash_map<absl::string_view, std::pair<CommandSyntaxChecker::CommandSyntaxFlags,
-                                                       FormatterProviderCreateFunc>>;
-  static const FormatterProviderLookupTbl& getKnownFormatters();
 };
 
 /**
@@ -120,14 +36,6 @@ private:
 class SubstitutionFormatUtils {
 public:
   static FormatterPtr defaultSubstitutionFormatter();
-  // Optional references are not supported, but this method has large performance
-  // impact, so using reference_wrapper.
-  static const absl::optional<std::reference_wrapper<const std::string>>
-  protocolToString(const absl::optional<Http::Protocol>& protocol);
-  static const std::string&
-  protocolToStringOrDefault(const absl::optional<Http::Protocol>& protocol);
-  static const absl::optional<std::string> getHostname();
-  static const std::string getHostnameOrDefault();
 
 private:
   SubstitutionFormatUtils();
@@ -261,447 +169,6 @@ public:
 
 private:
   const StructFormatter struct_formatter_;
-};
-
-/**
- * FormatterProvider for string literals. It ignores headers and stream info and returns string by
- * which it was initialized.
- */
-class PlainStringFormatter : public FormatterProvider {
-public:
-  PlainStringFormatter(const std::string& str);
-
-  // FormatterProvider
-  absl::optional<std::string> format(const Http::RequestHeaderMap&, const Http::ResponseHeaderMap&,
-                                     const Http::ResponseTrailerMap&, const StreamInfo::StreamInfo&,
-                                     absl::string_view, AccessLog::AccessLogType) const override;
-  ProtobufWkt::Value formatValue(const Http::RequestHeaderMap&, const Http::ResponseHeaderMap&,
-                                 const Http::ResponseTrailerMap&, const StreamInfo::StreamInfo&,
-                                 absl::string_view, AccessLog::AccessLogType) const override;
-
-private:
-  ProtobufWkt::Value str_;
-};
-
-/**
- * FormatterProvider for numbers.
- */
-class PlainNumberFormatter : public FormatterProvider {
-public:
-  PlainNumberFormatter(double num);
-
-  // FormatterProvider
-  absl::optional<std::string> format(const Http::RequestHeaderMap&, const Http::ResponseHeaderMap&,
-                                     const Http::ResponseTrailerMap&, const StreamInfo::StreamInfo&,
-                                     absl::string_view, AccessLog::AccessLogType) const override;
-  ProtobufWkt::Value formatValue(const Http::RequestHeaderMap&, const Http::ResponseHeaderMap&,
-                                 const Http::ResponseTrailerMap&, const StreamInfo::StreamInfo&,
-                                 absl::string_view, AccessLog::AccessLogType) const override;
-
-private:
-  ProtobufWkt::Value num_;
-};
-
-/**
- * FormatterProvider for local_reply_body. It returns the string from `local_reply_body` argument.
- */
-class LocalReplyBodyFormatter : public FormatterProvider {
-public:
-  LocalReplyBodyFormatter() = default;
-
-  // Formatter::format
-  absl::optional<std::string> format(const Http::RequestHeaderMap&, const Http::ResponseHeaderMap&,
-                                     const Http::ResponseTrailerMap&, const StreamInfo::StreamInfo&,
-                                     absl::string_view local_reply_body,
-                                     AccessLog::AccessLogType) const override;
-  ProtobufWkt::Value formatValue(const Http::RequestHeaderMap&, const Http::ResponseHeaderMap&,
-                                 const Http::ResponseTrailerMap&, const StreamInfo::StreamInfo&,
-                                 absl::string_view local_reply_body,
-                                 AccessLog::AccessLogType) const override;
-};
-
-/**
- * FormatterProvider for access log type. It returns the string from `access_log_type` argument.
- */
-class AccessLogTypeFormatter : public FormatterProvider {
-public:
-  AccessLogTypeFormatter() = default;
-
-  // Formatter::format
-  absl::optional<std::string> format(const Http::RequestHeaderMap&, const Http::ResponseHeaderMap&,
-                                     const Http::ResponseTrailerMap&, const StreamInfo::StreamInfo&,
-                                     absl::string_view local_reply_body,
-                                     AccessLog::AccessLogType access_log_type) const override;
-  ProtobufWkt::Value formatValue(const Http::RequestHeaderMap&, const Http::ResponseHeaderMap&,
-                                 const Http::ResponseTrailerMap&, const StreamInfo::StreamInfo&,
-                                 absl::string_view local_reply_body,
-                                 AccessLog::AccessLogType access_log_type) const override;
-};
-
-class HeaderFormatter {
-public:
-  HeaderFormatter(const std::string& main_header, const std::string& alternative_header,
-                  absl::optional<size_t> max_length);
-
-protected:
-  absl::optional<std::string> format(const Http::HeaderMap& headers) const;
-  ProtobufWkt::Value formatValue(const Http::HeaderMap& headers) const;
-
-private:
-  const Http::HeaderEntry* findHeader(const Http::HeaderMap& headers) const;
-
-  Http::LowerCaseString main_header_;
-  Http::LowerCaseString alternative_header_;
-  absl::optional<size_t> max_length_;
-};
-
-/**
- * FormatterProvider for headers byte size.
- */
-class HeadersByteSizeFormatter : public FormatterProvider {
-public:
-  // TODO(taoxuy): Add RequestTrailers here.
-  enum class HeaderType { RequestHeaders, ResponseHeaders, ResponseTrailers };
-
-  HeadersByteSizeFormatter(const HeaderType header_type);
-
-  absl::optional<std::string> format(const Http::RequestHeaderMap& request_headers,
-                                     const Http::ResponseHeaderMap& response_headers,
-                                     const Http::ResponseTrailerMap& response_trailers,
-                                     const StreamInfo::StreamInfo&, absl::string_view,
-                                     AccessLog::AccessLogType) const override;
-  ProtobufWkt::Value formatValue(const Http::RequestHeaderMap& request_headers,
-                                 const Http::ResponseHeaderMap& response_headers,
-                                 const Http::ResponseTrailerMap& response_trailers,
-                                 const StreamInfo::StreamInfo&, absl::string_view,
-                                 AccessLog::AccessLogType) const override;
-
-private:
-  uint64_t extractHeadersByteSize(const Http::RequestHeaderMap& request_headers,
-                                  const Http::ResponseHeaderMap& response_headers,
-                                  const Http::ResponseTrailerMap& response_trailers) const;
-  HeaderType header_type_;
-};
-
-/**
- * FormatterProvider for request headers.
- */
-class RequestHeaderFormatter : public FormatterProvider, HeaderFormatter {
-public:
-  RequestHeaderFormatter(const std::string& main_header, const std::string& alternative_header,
-                         absl::optional<size_t> max_length);
-
-  // FormatterProvider
-  absl::optional<std::string> format(const Http::RequestHeaderMap& request_headers,
-                                     const Http::ResponseHeaderMap&,
-                                     const Http::ResponseTrailerMap&, const StreamInfo::StreamInfo&,
-                                     absl::string_view, AccessLog::AccessLogType) const override;
-  ProtobufWkt::Value formatValue(const Http::RequestHeaderMap&, const Http::ResponseHeaderMap&,
-                                 const Http::ResponseTrailerMap&, const StreamInfo::StreamInfo&,
-                                 absl::string_view, AccessLog::AccessLogType) const override;
-};
-
-/**
- * FormatterProvider for response headers.
- */
-class ResponseHeaderFormatter : public FormatterProvider, HeaderFormatter {
-public:
-  ResponseHeaderFormatter(const std::string& main_header, const std::string& alternative_header,
-                          absl::optional<size_t> max_length);
-
-  // FormatterProvider
-  absl::optional<std::string> format(const Http::RequestHeaderMap&,
-                                     const Http::ResponseHeaderMap& response_headers,
-                                     const Http::ResponseTrailerMap&, const StreamInfo::StreamInfo&,
-                                     absl::string_view, AccessLog::AccessLogType) const override;
-  ProtobufWkt::Value formatValue(const Http::RequestHeaderMap&, const Http::ResponseHeaderMap&,
-                                 const Http::ResponseTrailerMap&, const StreamInfo::StreamInfo&,
-                                 absl::string_view, AccessLog::AccessLogType) const override;
-};
-
-/**
- * FormatterProvider for response trailers.
- */
-class ResponseTrailerFormatter : public FormatterProvider, HeaderFormatter {
-public:
-  ResponseTrailerFormatter(const std::string& main_header, const std::string& alternative_header,
-                           absl::optional<size_t> max_length);
-
-  // FormatterProvider
-  absl::optional<std::string> format(const Http::RequestHeaderMap&, const Http::ResponseHeaderMap&,
-                                     const Http::ResponseTrailerMap& response_trailers,
-                                     const StreamInfo::StreamInfo&, absl::string_view,
-                                     AccessLog::AccessLogType) const override;
-  ProtobufWkt::Value formatValue(const Http::RequestHeaderMap&, const Http::ResponseHeaderMap&,
-                                 const Http::ResponseTrailerMap&, const StreamInfo::StreamInfo&,
-                                 absl::string_view, AccessLog::AccessLogType) const override;
-};
-
-class GrpcStatusFormatter : public FormatterProvider, HeaderFormatter {
-public:
-  enum Format {
-    CamelString,
-    SnakeString,
-    Number,
-  };
-
-  GrpcStatusFormatter(const std::string& main_header, const std::string& alternative_header,
-                      absl::optional<size_t> max_length, Format format);
-
-  // FormatterProvider
-  absl::optional<std::string> format(const Http::RequestHeaderMap&,
-                                     const Http::ResponseHeaderMap& response_headers,
-                                     const Http::ResponseTrailerMap& response_trailers,
-                                     const StreamInfo::StreamInfo&, absl::string_view,
-                                     AccessLog::AccessLogType) const override;
-  ProtobufWkt::Value formatValue(const Http::RequestHeaderMap&, const Http::ResponseHeaderMap&,
-                                 const Http::ResponseTrailerMap&, const StreamInfo::StreamInfo&,
-                                 absl::string_view, AccessLog::AccessLogType) const override;
-
-  static Format parseFormat(absl::string_view format);
-
-private:
-  const Format format_;
-};
-
-/**
- * FormatterProvider based on StreamInfo fields.
- */
-class StreamInfoFormatter : public FormatterProvider {
-public:
-  StreamInfoFormatter(const std::string&, const std::string& = "",
-                      const absl::optional<size_t>& = absl::nullopt);
-
-  // FormatterProvider
-  absl::optional<std::string> format(const Http::RequestHeaderMap&, const Http::ResponseHeaderMap&,
-                                     const Http::ResponseTrailerMap&, const StreamInfo::StreamInfo&,
-                                     absl::string_view, AccessLog::AccessLogType) const override;
-  ProtobufWkt::Value formatValue(const Http::RequestHeaderMap&, const Http::ResponseHeaderMap&,
-                                 const Http::ResponseTrailerMap&, const StreamInfo::StreamInfo&,
-                                 absl::string_view, AccessLog::AccessLogType) const override;
-
-  class FieldExtractor {
-  public:
-    virtual ~FieldExtractor() = default;
-
-    virtual absl::optional<std::string> extract(const StreamInfo::StreamInfo&) const PURE;
-    virtual ProtobufWkt::Value extractValue(const StreamInfo::StreamInfo&) const PURE;
-  };
-  using FieldExtractorPtr = std::unique_ptr<FieldExtractor>;
-  using FieldExtractorCreateFunc =
-      std::function<FieldExtractorPtr(const std::string&, const absl::optional<size_t>&)>;
-
-  enum class StreamInfoAddressFieldExtractionType { WithPort, WithoutPort, JustPort };
-  StreamInfoFormatter(FieldExtractorPtr field_extractor) {
-    field_extractor_ = std::move(field_extractor);
-  }
-
-private:
-  FieldExtractorPtr field_extractor_;
-
-  using FieldExtractorLookupTbl =
-      absl::flat_hash_map<absl::string_view,
-                          std::pair<CommandSyntaxChecker::CommandSyntaxFlags,
-                                    StreamInfoFormatter::FieldExtractorCreateFunc>>;
-  static const FieldExtractorLookupTbl& getKnownFieldExtractors();
-};
-
-/**
- * Base formatter for formatting Metadata objects
- */
-
-class MetadataFormatter : public FormatterProvider {
-public:
-  using GetMetadataFunction =
-      std::function<const envoy::config::core::v3::Metadata*(const StreamInfo::StreamInfo&)>;
-  MetadataFormatter(const std::string& filter_namespace, const std::vector<std::string>& path,
-                    absl::optional<size_t> max_length, GetMetadataFunction get);
-
-  absl::optional<std::string> format(const Http::RequestHeaderMap&, const Http::ResponseHeaderMap&,
-                                     const Http::ResponseTrailerMap&,
-                                     const StreamInfo::StreamInfo& stream_info, absl::string_view,
-                                     AccessLog::AccessLogType) const override;
-
-  ProtobufWkt::Value formatValue(const Http::RequestHeaderMap&, const Http::ResponseHeaderMap&,
-                                 const Http::ResponseTrailerMap&,
-                                 const StreamInfo::StreamInfo& stream_info, absl::string_view,
-                                 AccessLog::AccessLogType) const override;
-
-protected:
-  absl::optional<std::string>
-  formatMetadata(const envoy::config::core::v3::Metadata& metadata) const;
-  ProtobufWkt::Value formatMetadataValue(const envoy::config::core::v3::Metadata& metadata) const;
-
-private:
-  std::string filter_namespace_;
-  std::vector<std::string> path_;
-  absl::optional<size_t> max_length_;
-  GetMetadataFunction get_func_;
-};
-
-/**
- * FormatterProvider for DynamicMetadata from StreamInfo.
- */
-class DynamicMetadataFormatter : public MetadataFormatter {
-public:
-  DynamicMetadataFormatter(const std::string& filter_namespace,
-                           const std::vector<std::string>& path, absl::optional<size_t> max_length);
-};
-
-/**
- * FormatterProvider for ClusterMetadata from StreamInfo.
- */
-class ClusterMetadataFormatter : public MetadataFormatter {
-public:
-  ClusterMetadataFormatter(const std::string& filter_namespace,
-                           const std::vector<std::string>& path, absl::optional<size_t> max_length);
-};
-
-/**
- * FormatterProvider for UpstreamHostMetadata from StreamInfo.
- */
-class UpstreamHostMetadataFormatter : public MetadataFormatter {
-public:
-  UpstreamHostMetadataFormatter(const std::string& filter_namespace,
-                                const std::vector<std::string>& path,
-                                absl::optional<size_t> max_length);
-};
-
-/**
- * FormatterProvider for FilterState from StreamInfo.
- */
-class FilterStateFormatter : public FormatterProvider {
-public:
-  static std::unique_ptr<FilterStateFormatter>
-  create(const std::string& format, const absl::optional<size_t>& max_length, bool is_upstream);
-
-  FilterStateFormatter(const std::string& key, absl::optional<size_t> max_length,
-                       bool serialize_as_string, bool is_upstream = false);
-
-  // FormatterProvider
-  absl::optional<std::string> format(const Http::RequestHeaderMap&, const Http::ResponseHeaderMap&,
-                                     const Http::ResponseTrailerMap&, const StreamInfo::StreamInfo&,
-                                     absl::string_view, AccessLog::AccessLogType) const override;
-  ProtobufWkt::Value formatValue(const Http::RequestHeaderMap&, const Http::ResponseHeaderMap&,
-                                 const Http::ResponseTrailerMap&, const StreamInfo::StreamInfo&,
-                                 absl::string_view, AccessLog::AccessLogType) const override;
-
-private:
-  const Envoy::StreamInfo::FilterState::Object*
-  filterState(const StreamInfo::StreamInfo& stream_info) const;
-
-  std::string key_;
-  absl::optional<size_t> max_length_;
-
-  bool serialize_as_string_;
-  const bool is_upstream_;
-};
-
-/**
- * Base FormatterProvider for system times from StreamInfo.
- */
-class SystemTimeFormatter : public FormatterProvider {
-public:
-  using TimeFieldExtractor =
-      std::function<absl::optional<SystemTime>(const StreamInfo::StreamInfo& stream_info)>;
-  using TimeFieldExtractorPtr = std::unique_ptr<TimeFieldExtractor>;
-
-  SystemTimeFormatter(const std::string& format, TimeFieldExtractorPtr f);
-
-  // FormatterProvider
-  absl::optional<std::string> format(const Http::RequestHeaderMap&, const Http::ResponseHeaderMap&,
-                                     const Http::ResponseTrailerMap&, const StreamInfo::StreamInfo&,
-                                     absl::string_view, AccessLog::AccessLogType) const override;
-  ProtobufWkt::Value formatValue(const Http::RequestHeaderMap&, const Http::ResponseHeaderMap&,
-                                 const Http::ResponseTrailerMap&, const StreamInfo::StreamInfo&,
-                                 absl::string_view, AccessLog::AccessLogType) const override;
-
-private:
-  const Envoy::DateFormatter date_formatter_;
-  const TimeFieldExtractorPtr time_field_extractor_;
-};
-
-/**
- * SystemTimeFormatter (FormatterProvider) for request start time from StreamInfo.
- */
-class StartTimeFormatter : public SystemTimeFormatter {
-public:
-  StartTimeFormatter(const std::string& format);
-};
-
-/**
- * SystemTimeFormatter (FormatterProvider) for downstream cert start time from the StreamInfo's
- * ConnectionInfo.
- */
-class DownstreamPeerCertVStartFormatter : public SystemTimeFormatter {
-public:
-  DownstreamPeerCertVStartFormatter(const std::string& format);
-};
-
-/**
- * SystemTimeFormatter (FormatterProvider) for downstream cert end time from the StreamInfo's
- * ConnectionInfo.
- */
-class DownstreamPeerCertVEndFormatter : public SystemTimeFormatter {
-public:
-  DownstreamPeerCertVEndFormatter(const std::string& format);
-};
-
-/**
- * SystemTimeFormatter (FormatterProvider) for upstream cert start time from the StreamInfo's
- * upstreamInfo.
- */
-class UpstreamPeerCertVStartFormatter : public SystemTimeFormatter {
-public:
-  UpstreamPeerCertVStartFormatter(const std::string& format);
-};
-
-/**
- * SystemTimeFormatter (FormatterProvider) for upstream cert end time from the StreamInfo's
- * upstreamInfo.
- */
-class UpstreamPeerCertVEndFormatter : public SystemTimeFormatter {
-public:
-  UpstreamPeerCertVEndFormatter(const std::string& format);
-};
-
-/**
- * FormatterProvider for environment. If no valid environment value then
- */
-class EnvironmentFormatter : public FormatterProvider {
-public:
-  EnvironmentFormatter(const std::string& key, absl::optional<size_t> max_length);
-
-  // FormatterProvider
-  absl::optional<std::string> format(const Http::RequestHeaderMap&, const Http::ResponseHeaderMap&,
-                                     const Http::ResponseTrailerMap&, const StreamInfo::StreamInfo&,
-                                     absl::string_view, AccessLog::AccessLogType) const override;
-  ProtobufWkt::Value formatValue(const Http::RequestHeaderMap&, const Http::ResponseHeaderMap&,
-                                 const Http::ResponseTrailerMap&, const StreamInfo::StreamInfo&,
-                                 absl::string_view, AccessLog::AccessLogType) const override;
-
-private:
-  ProtobufWkt::Value str_;
-};
-
-/**
- * FormatterProvider for request headers from StreamInfo (rather than the request_headers param).
- * Purely for testing.
- */
-class StreamInfoRequestHeaderFormatter : public FormatterProvider, HeaderFormatter {
-public:
-  StreamInfoRequestHeaderFormatter(const std::string& main_header,
-                                   const std::string& alternative_header,
-                                   absl::optional<size_t> max_length);
-
-  // FormatterProvider
-  absl::optional<std::string> format(const Http::RequestHeaderMap& request_headers,
-                                     const Http::ResponseHeaderMap&,
-                                     const Http::ResponseTrailerMap&, const StreamInfo::StreamInfo&,
-                                     absl::string_view, AccessLog::AccessLogType) const override;
-  ProtobufWkt::Value formatValue(const Http::RequestHeaderMap&, const Http::ResponseHeaderMap&,
-                                 const Http::ResponseTrailerMap&, const StreamInfo::StreamInfo&,
-                                 absl::string_view, AccessLog::AccessLogType) const override;
 };
 
 } // namespace Formatter
