+++ tools/code_format/config.yaml
@@ -116,6 +116,8 @@ paths:
     - source/common/network/lc_trie.h
     - source/common/network/socket_impl.cc
     - source/common/ssl/tls_certificate_config_impl.cc
+    - source/common/formatter/http_specific_formatter.cc
+    - source/common/formatter/stream_info_formatter.cc
     - source/common/formatter/substitution_formatter.cc
     - source/common/formatter/substitution_format_string.cc
     - source/common/stats/tag_extractor_impl.cc
@@ -310,6 +312,8 @@ paths:
     - source/common/common/regex.h
     - source/common/common/regex.cc
     - source/common/common/utility.cc
+    - source/common/formatter/http_specific_formatter.cc
+    - source/common/formatter/stream_info_formatter.cc
     - source/common/formatter/substitution_formatter.cc
     - source/common/stats/tag_extractor_impl.h
     - source/common/stats/tag_extractor_impl.cc
